pyproject.toml
[tool.poetry]
name = "btmonitor"
version = "0.1.0"
description = "Real-time bus traffic monitor (server). Includes scraper for monitoring bus traffing in Tallinn, Estonia."
authors = ["Toomas Uudisaru <toomas.uudisaru@gmail.com>"]
license = "MIT license"

readme = "README.rst"

repository = "https://github.com/uudisaru/btmonitor"

homepage = "https://uudisaru.github.io/btmonitor"

packages = [
    { include = "btmonitor", from = "src"},
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.7",
]
[tool.poetry.dependencies]
python = "^3.7"
# List project dependencies here
aiohttp = "^3.5"
click = "^7.0"
uvloop = "^0.12.2"
sanic = "^19.6"
msgpack = "^0.6.1"
pyjwt = "^1.7"
cryptography = "^2.7"

[tool.poetry.dev-dependencies]
bump2version = "^0.5.0"
pip = "^18.1"
wheel = "^0.32.1"
watchdog = "^0.9.0"
flake8 = "^3.5.0"
tox = "^3.5.2"
coverage = "^4.5.1"
Sphinx = "^1.8.1"
twine = "^1.12.1"

pytest = "^3.8.2"
pytest-runner = "^4.2"
pytest-cov = "^2.6.0"
setuptools = "^40.5.0"
rstcheck = "^3.3.1"
ptpython = "^2.0.4"
pytest-asyncio = "^0.10.0"

[tool.poetry.scripts]
btmonitor = "btmonitor.main:main"
